+++ BraintreePaymentFlow/BTPaymentFlowDriver.m
@@ -83,6 +83,7 @@ static BTPaymentFlowDriver *paymentFlowDriver;
     if ([self.viewControllerPresentingDelegate respondsToSelector:@selector(paymentDriver:requestsPresentationOfViewController:)]) {
         self.safariViewController = [[SFSafariViewController alloc] initWithURL:appSwitchURL];
         self.safariViewController.delegate = self;
+        self.safariViewController.dismissButtonStyle = SFSafariViewControllerDismissButtonStyleCancel;
         [self.viewControllerPresentingDelegate paymentDriver:self requestsPresentationOfViewController:self.safariViewController];
     } else {
         [[BTLogger sharedLogger] critical:@"Unable to display View Controller to continue payment flow. BTPaymentFlowDriver needs a viewControllerPresentingDelegate<BTViewControllerPresentingDelegate> to be set."];