+++ BraintreePaymentFlowTests/BraintreePaymentFlowTests-Bridging-Header.h
@@ -1,0 +1,1 @@
+#import <BraintreePaymentFlow/BTPaymentFlowDriver_Internal.h>